fix:#9 cache 갱신 코드 admin으로 도메인 변경
